requirements.txt
requests
tinker
python-dotenv
